requirements.txt
anthropic>=0.40.0
openai>=1.0.0
langgraph>=0.2.0
langchain-core>=0.1
tavily-python>=0.3.0
fastapi>=0.110.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
pyyaml>=6.0
streamlit>=1.12.0
chainlit>=1.0.0
httpx>=0.25.0
pydantic>=2.6.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0
sse-starlette>=1.8.0
